This is the code I used to test the algorithm
Some of the functions have been updated from the Arduino file others are just for testing. Run this in a c++ IDE (i used DEV C++). It pretends to look at sensors and pulls the data from a modelMaze. There are 4 mazes coded in. In the main file at the bottom of the file uncomment one of the initializeModelMaze function calls that will be the maze that it maps. In initialize maze(line 606), you can change the starting position of the maze, make sure to also update goTo(&maze[][]); (line 1964) with the same coordinate. If you want to code your own maze and add it let me know and i can show you how. If you find any bugs let me know and tell me what maze and where you started that way i know hoe to recreate it.